Pushing to a merge request clears the approvals list if the respective project setting is enabled
Previously if the project had approvals_before_merge setting at 0
the approvals would not be cleared even if the "Pushing to a merge
request clears approvals" project setting was enabled.

This commit removes the approvals_before_merge out of the check of
whether we should or not reset the merge request approvals.

+++ ee/app/services/ee/merge_requests/update_service.rb
@@ -38,9 +38,7 @@ module EE
       def reset_approvals(merge_request)
         target_project = merge_request.target_project
 
-        if target_project.approvals_before_merge.nonzero? && target_project.reset_approvals_on_push
-          merge_request.approvals.delete_all
-        end
+        merge_request.approvals.delete_all if target_project.reset_approvals_on_push
       end
     end
   end
